WIP: Split ImageElements gatherer for snapshot mode
