+++ lighthouse-core/fraggle-rock/config/default-config.js
@@ -29,6 +29,7 @@ const artifacts = {
   HostUserAgent: '',
   IFrameElements: '',
   ImageElements: '',
+  ImageElementsSnapshot: '',
   InstallabilityErrors: '',
   InspectorIssues: '',
   JsUsage: '',
@@ -81,6 +82,7 @@ const defaultConfig = {
     {id: artifacts.GlobalListeners, gatherer: 'global-listeners'},
     {id: artifacts.IFrameElements, gatherer: 'iframe-elements'},
     {id: artifacts.ImageElements, gatherer: 'image-elements'},
+    {id: artifacts.ImageElementsSnapshot, gatherer: 'image-elements-snapshot'},
     {id: artifacts.InstallabilityErrors, gatherer: 'installability-errors'},
     {id: artifacts.InspectorIssues, gatherer: 'inspector-issues'},
     {id: artifacts.JsUsage, gatherer: 'js-usage'},
